Update from v1.4.57 tot v1.5.1 (#85)
* Update from v1.4.57 tot v1.5.1

* Removed Sonarcloud badge

+++ src/Domain/Publication/Dossier/Type/WooDecision/WooDecisionDeleteStrategy.php
@@ -5,35 +5,36 @@ declare(strict_types=1);
 namespace App\Domain\Publication\Dossier\Type\WooDecision;
 
 use App\Domain\Publication\Dossier\AbstractDossier;
-use App\Domain\Publication\Dossier\DossierDeleteHelper;
-use App\Domain\Publication\Dossier\Type\DossierDeleteStrategyInterface;
+use App\Domain\Publication\Dossier\AbstractEntityWithFileInfoDeleteStrategy;
 use App\Service\BatchDownloadService;
 use App\Service\DocumentService;
 use App\Service\Inquiry\InquiryService;
-use Webmozart\Assert\Assert;
+use App\Service\Storage\DocumentStorageService;
 
-readonly class WooDecisionDeleteStrategy implements DossierDeleteStrategyInterface
+readonly class WooDecisionDeleteStrategy extends AbstractEntityWithFileInfoDeleteStrategy
 {
     public function __construct(
-        private DossierDeleteHelper $dossierDeleteHelper,
+        DocumentStorageService $storageService,
         private DocumentService $documentService,
         private BatchDownloadService $downloadService,
         private InquiryService $inquiryService,
     ) {
+        parent::__construct($storageService);
     }
 
     public function delete(AbstractDossier $dossier): void
     {
-        Assert::isInstanceOf($dossier, WooDecision::class);
+        if (! $dossier instanceof WooDecision) {
+            return;
+        }
 
         foreach ($dossier->getDocuments() as $document) {
             $this->documentService->removeDocumentFromDossier($dossier, $document, false);
         }
 
-        $this->dossierDeleteHelper->deleteFileForEntity($dossier->getInventory());
-        $this->dossierDeleteHelper->deleteFileForEntity($dossier->getRawInventory());
-        $this->dossierDeleteHelper->deleteFileForEntity($dossier->getDecisionDocument());
-        $this->dossierDeleteHelper->deleteAttachments($dossier->getAttachments());
+        $this->deleteFileForEntity($dossier->getInventory());
+        $this->deleteFileForEntity($dossier->getRawInventory());
+        $this->deleteFileForEntity($dossier->getDecisionDocument());
 
         $this->downloadService->removeAllDownloadsForEntity($dossier);
         $this->inquiryService->removeDossierFromInquiries($dossier);
